:art: refactor multiple classes
