+++ staff.rb
@@ -1,0 +1,49 @@
+class Staff 
+	attr_reader :age, :salary, :phase, :target_raise, :response
+  	attr_accessor :name
+
+  def initialize(options={})
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+    @phase = 3
+  end
+
+  def offer_high_five
+    "High five!"
+  end
+
+  def teach_stuff
+    self.response += "Listen, class, this is how everything works. "
+    self.response += "*drops crazy knowledge bomb* "
+    self.response += "... You're welcome."
+    self.response
+  end
+
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+   def set_performance_rating(rating)
+    response = ""
+    if rating > self.class::@rating_rank
+      receive_raise(@target_raise)
+      response = "Yay, I'm a great employee!"
+    else
+      response += "Oh, well -- thanks to this actionable, specific, and kind "
+      response += "feedback, I'll do better next time."
+    end
+    response
+  end
+  
+end
+
